+++ config/event-projector.php
@@ -3,21 +3,29 @@
 return [
 
     /*
-     * Projectors are classes that build up projections. You can create them by
-     * performing `php artisan event-projector:create-projector`. Projectors
-     * can be registered in this array or a service provider.
+     * These directories will be scanned for Projectors and Reactors. They
+     * will be registered to Projectionist automatically.
+     */
+    'auto_discover_projectors_and_reactors' => [
+        app_path(),
+    ],
+
+    /*
+     * Projectors are classes that build up projections. You can create them by performing
+     * `php artisan event-projector:create-projector`. When not using auto-discovery,
+     * Projectors can be registered in this array or a service provider.
      */
     'projectors' => [
         // App\Projectors\YourProjector::class
     ],
 
     /*
-     * Reactors are classes that handle side effects. You can create them by
-     * performing `php artisan event-projector:create-reactor`. Reactors
-     * can be registered in this array or a service provider.
+     * Reactors are classes that handle side-effects. You can create them by performing
+     * `php artisan event-projector:create-reactor`. When not using auto-discovery
+     * Reactors can be registered in this array or a service provider.
      */
     'reactors' => [
-        // App\Reactors\YourReactors::class
+        // App\Reactors\YourReactor::class
     ],
 
     /*
@@ -27,9 +35,9 @@ return [
     'queue' => env('EVENT_PROJECTOR_QUEUE_NAME', null),
 
     /*
-     * When a projector or reactor throws an exception the event projectionist can catch it
+     * When a Projector or Reactor throws an exception the event Projectionist can catch it
      * so all other projectors and reactors can still do their work. The exception will
-     * be passed to the `handleException` method on that projector or reactor.
+     * be passed to the `handleException` method on that Projector or Reactor.
      */
     'catch_exceptions' => env('EVENT_PROJECTOR_CATCH_EXCEPTIONS', false),
 
@@ -41,16 +49,9 @@ return [
     'stored_event_model' => \Spatie\EventProjector\Models\StoredEvent::class,
 
     /*
-     * This class is responsible for projector statuses. To add extra behaviour you
+     * This class is responsible for handling stored events. To add extra behaviour you
      * can change this to a class of your own. The only restriction is that
-     * it should extend \Spatie\EventProjector\Models\ProjectorStatus.
-     */
-    'projector_status_model' => \Spatie\EventProjector\Models\ProjectorStatus::class,
-
-    /*
-     * This class is responsible for handle stored events. To add extra behaviour you
-     * can change this to a class of your own. The only restriction is that
-     * it should extend \Spatie\EventProjector\HandleStoredEventJob.
+     * it should extend \Spatie\EventProjector\HandleDomainEventJob.
      */
     'stored_event_job' => \Spatie\EventProjector\HandleStoredEventJob::class,
 
@@ -62,9 +63,18 @@ return [
     'event_serializer' => \Spatie\EventProjector\EventSerializers\JsonEventSerializer::class,
 
     /*
-     * When replaying events potentially a lot of events will have to be retrieved.
-     * In order to avoid memory problems events will be retrieved in
-     * a chuncked way. You can specify the chunk size here.
+     * When replaying events, potentially a lot of events will have to be retrieved.
+     * In order to avoid memory problems events will be retrieved as chunks.
+     * You can specify the chunk size here.
      */
     'replay_chunk_size' => 1000,
+
+    /*
+     * In production, you likely don't want the package to auto-discover the event handlers
+     * on every request. The package can cache all registered event handlers.
+     * More info: https://docs.spatie.be/laravel-event-projector/v2/advanced-usage/discovering-projectors-and-reactors
+     *
+     * Here you can specify where the cache should be stored.
+     */
+    'cache_path' => storage_path('app/event-projector'),
 ];