+++ src/EventHandlers/HandlesEvents.php
@@ -3,7 +3,11 @@
 namespace Spatie\EventProjector\EventHandlers;
 
 use Exception;
+use ReflectionClass;
+use ReflectionMethod;
+use ReflectionParameter;
 use Illuminate\Support\Collection;
+use Spatie\EventProjector\ShouldBeStored;
 use Spatie\EventProjector\Models\StoredEvent;
 use Spatie\EventProjector\Exceptions\InvalidEventHandler;
 
@@ -21,8 +25,9 @@ trait HandlesEvents
         $handlerClassOrMethod = $this->getEventHandlingMethods()->get($eventClass);
 
         $parameters = [
-            'event'       => $storedEvent->event,
+            'event' => $storedEvent->event,
             'storedEvent' => $storedEvent,
+            'aggregateUuid' => $storedEvent->aggregate_uuid,
         ];
 
         if (class_exists($handlerClassOrMethod)) {
@@ -36,14 +41,14 @@ trait HandlesEvents
         app()->call([$this, $handlerClassOrMethod], $parameters);
     }
 
-    public function handleException(Exception $exception)
+    public function handleException(Exception $exception): void
     {
         report($exception);
     }
 
-    protected function getEventHandlingMethods(): Collection
+    public function getEventHandlingMethods(): Collection
     {
-        return collect($this->handlesEvents ?? [])
+        $handlesEvents = collect($this->handlesEvents ?? [])
             ->mapWithKeys(function (string $handlerMethod, $eventClass) {
                 if (is_numeric($eventClass)) {
                     return [$handlerMethod => 'on'.ucfirst(class_basename($handlerMethod))];
@@ -51,5 +56,38 @@ trait HandlesEvents
 
                 return [$eventClass => $handlerMethod];
             });
+
+        if ($this->handleEvent ?? false) {
+            $handlesEvents->put($this->handleEvent, get_class($this));
+        }
+
+        if (! isset($this->handlesEvents) && ! isset($this->handleEvent)) {
+            $handlesEvents = $this->autoDetectHandlesEvents();
+        }
+
+        return $handlesEvents;
+    }
+
+    private function autoDetectHandlesEvents(): Collection
+    {
+        return collect((new ReflectionClass($this))->getMethods())
+            ->flatMap(function (ReflectionMethod $method) {
+                $method = new ReflectionMethod($this, $method->name);
+
+                $eventClass = collect($method->getParameters())
+                    ->map(function (ReflectionParameter $parameter) {
+                        return optional($parameter->getType())->getName();
+                    })
+                    ->first(function ($typeHint) {
+                        return is_subclass_of($typeHint, ShouldBeStored::class);
+                    });
+
+                if (! $eventClass) {
+                    return;
+                }
+
+                return [$eventClass => $method->name];
+            })
+            ->filter();
     }
 }