+++ src/Exceptions/InvalidEventHandler.php
@@ -5,9 +5,9 @@ namespace Spatie\EventProjector\Exceptions;
 use Exception;
 use Spatie\EventProjector\ShouldBeStored;
 
-class InvalidEventHandler extends Exception
+final class InvalidEventHandler extends Exception
 {
-    public static function eventHandlingMethodDoesNotExist(object $eventHandler, ShouldBeStored $event, string $methodName)
+    public static function eventHandlingMethodDoesNotExist(object $eventHandler, ShouldBeStored $event, string $methodName): self
     {
         $eventHandlerClass = get_class($eventHandler);
         $eventClass = get_class($event);
@@ -28,5 +28,10 @@ class InvalidEventHandler extends Exception
     public static function notAnEventHandler(object $object)
     {
         return new static('`'.get_class($object).'` must implement Spatie\EventProjector\EventHandlers\EventHandler');
+    }
+
+    public static function notAnEventHandlingClassName(string $className)
+    {
+        return new static('`'.$className.'` must implement Spatie\EventProjector\EventHandlers\EventHandler');
     }
 }