+++ src/Projectionist.php
@@ -3,41 +3,49 @@
 namespace Spatie\EventProjector;
 
 use Exception;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Spatie\EventProjector\Models\StoredEvent;
 use Spatie\EventProjector\Projectors\Projector;
 use Spatie\EventProjector\EventHandlers\EventHandler;
 use Spatie\EventProjector\Events\FinishedEventReplay;
 use Spatie\EventProjector\Events\StartingEventReplay;
+use Spatie\EventProjector\Projectors\QueuedProjector;
 use Spatie\EventProjector\Exceptions\InvalidEventHandler;
 use Spatie\EventProjector\EventHandlers\EventHandlerCollection;
 use Spatie\EventProjector\Events\EventHandlerFailedHandlingEvent;
-use Spatie\EventProjector\Events\ProjectorDidNotHandlePriorEvents;
 
-class Projectionist
+final class Projectionist
 {
-    /** @var array */
-    protected $config;
+    /** @var \Spatie\EventProjector\EventHandlers\EventHandlerCollection */
+    private $projectors;
 
     /** @var \Spatie\EventProjector\EventHandlers\EventHandlerCollection */
-    protected $projectors;
-
-    /** @var \Spatie\EventProjector\EventHandlers\EventHandlerCollection */
-    protected $reactors;
+    private $reactors;
 
     /** @var bool */
-    protected $isProjecting = false;
+    private $catchExceptions;
 
     /** @var bool */
-    protected $isReplaying = false;
+    private $replayChunkSize;
 
-    public function __construct(array $config = [])
+    /** @var string */
+    private $storedEventClass;
+
+    /** @var bool */
+    private $isProjecting = false;
+
+    /** @var bool */
+    private $isReplaying = false;
+
+    public function __construct(array $config)
     {
         $this->projectors = new EventHandlerCollection();
-
         $this->reactors = new EventHandlerCollection();
 
-        $this->config = $config;
+        $this->catchExceptions = $config['catch_exceptions'];
+        $this->replayChunkSize = $config['replay_chunk_size'];
+        $this->storedEventClass = $config['stored_event_model'];
     }
 
     public function addProjector($projector): Projectionist
@@ -53,6 +61,29 @@ class Projectionist
         $this->projectors->add($projector);
 
         return $this;
+    }
+
+    public function withoutEventHandlers(array $eventHandlers = null): Projectionist
+    {
+        if (is_null($eventHandlers)) {
+            $this->projectors = new EventHandlerCollection();
+            $this->reactors = new EventHandlerCollection();
+
+            return $this;
+        }
+
+        $eventHandlers = Arr::wrap($eventHandlers);
+
+        $this->projectors->remove($eventHandlers);
+
+        $this->reactors->remove($eventHandlers);
+
+        return $this;
+    }
+
+    public function withoutEventHandler(string $eventHandler): Projectionist
+    {
+        return $this->withoutEventHandlers([$eventHandler]);
     }
 
     public function addProjectors(array $projectors): Projectionist
@@ -78,6 +109,14 @@ class Projectionist
 
     public function addReactor($reactor): Projectionist
     {
+        if (is_string($reactor)) {
+            $reactor = app($reactor);
+        }
+
+        if (! $reactor instanceof EventHandler) {
+            throw InvalidEventHandler::notAnEventHandler($reactor);
+        }
+
         $this->reactors->add($reactor);
 
         return $this;
@@ -97,26 +136,51 @@ class Projectionist
         return $this->reactors->all();
     }
 
-    public function storeEvent(ShouldBeStored $event)
+    public function addEventHandler($eventHandlerClass)
     {
-        $storedEvent = $this->getStoredEventClass()::createForEvent($event);
-
-        $this->handleImmediately($storedEvent);
-
-        if (method_exists($event, 'tags')) {
-            $tags = $event->tags();
+        if (! is_string($eventHandlerClass)) {
+            $eventHandlerClass = get_class($eventHandlerClass);
         }
 
-        $storedEventJob = $this->getStoredEventJob()::createForEvent($storedEvent, $tags ?? []);
+        if (is_subclass_of($eventHandlerClass, Projector::class)) {
+            $this->addProjector($eventHandlerClass);
 
-        dispatch($storedEventJob->onQueue($this->config['queue']));
+            return;
+        }
+
+        if (is_subclass_of($eventHandlerClass, QueuedProjector::class)) {
+            $this->addProjector($eventHandlerClass);
+
+            return;
+        }
+
+        if (is_subclass_of($eventHandlerClass, EventHandler::class)) {
+            $this->addReactor($eventHandlerClass);
+
+            return;
+        }
+
+        throw InvalidEventHandler::notAnEventHandlingClassName($eventHandlerClass);
     }
 
-    public function handle(StoredEvent $storedEvent)
+    public function addEventHandlers(array $eventHandlers)
     {
+        foreach ($eventHandlers as $eventHandler) {
+            $this->addEventHandler($eventHandler);
+        }
+    }
+
+    public function handle(StoredEvent $storedEvent): void
+    {
+        $projectors = $this->projectors
+            ->forEvent($storedEvent)
+            ->reject(function (Projector $projector) {
+                return $projector->shouldBeCalledImmediately();
+            });
+
         $this->applyStoredEventToProjectors(
             $storedEvent,
-            $this->projectors->forEvent($storedEvent)
+            $projectors
         );
 
         $this->applyStoredEventToReactors(
@@ -125,7 +189,7 @@ class Projectionist
         );
     }
 
-    public function handleImmediately(StoredEvent $storedEvent)
+    public function handleWithSyncProjectors(StoredEvent $storedEvent): void
     {
         $projectors = $this->projectors
             ->forEvent($storedEvent)
@@ -141,46 +205,30 @@ class Projectionist
         return $this->isProjecting;
     }
 
-    protected function applyStoredEventToProjectors(StoredEvent $storedEvent, Collection $projectors)
+    private function applyStoredEventToProjectors(StoredEvent $storedEvent, Collection $projectors): void
     {
         $this->isProjecting = true;
 
         foreach ($projectors as $projector) {
-            if ($projector->hasAlreadyReceivedEvent($storedEvent)) {
-                continue;
-            }
-
-            if (! $projector->hasReceivedAllPriorEvents($storedEvent)) {
-                event(new ProjectorDidNotHandlePriorEvents($projector, $storedEvent));
-
-                $projector->markAsNotUpToDate($storedEvent);
-
-                continue;
-            }
-
-            if (! $this->callEventHandler($projector, $storedEvent)) {
-                continue;
-            }
-
-            $projector->rememberReceivedEvent($storedEvent);
+            $this->callEventHandler($projector, $storedEvent);
         }
 
         $this->isProjecting = false;
     }
 
-    protected function applyStoredEventToReactors(StoredEvent $storedEvent, Collection $reactors)
+    private function applyStoredEventToReactors(StoredEvent $storedEvent, Collection $reactors): void
     {
         foreach ($reactors as $reactor) {
             $this->callEventHandler($reactor, $storedEvent);
         }
     }
 
-    protected function callEventHandler(EventHandler $eventHandler, StoredEvent $storedEvent): bool
+    private function callEventHandler(EventHandler $eventHandler, StoredEvent $storedEvent): bool
     {
         try {
             $eventHandler->handle($storedEvent);
         } catch (Exception $exception) {
-            if (! $this->config['catch_exceptions']) {
+            if (! $this->catchExceptions) {
                 throw $exception;
             }
 
@@ -199,19 +247,30 @@ class Projectionist
         return $this->isReplaying;
     }
 
-    public function replay(Collection $projectors, int $afterStoredEventId = 0, callable $onEventReplayed = null)
-    {
+    public function replay(
+        Collection $projectors,
+        int $startingFromEventId = 0,
+        callable $onEventReplayed = null
+    ): void {
+        $projectors = new EventHandlerCollection($projectors);
+
         $this->isReplaying = true;
 
-        $projectors = new EventHandlerCollection($projectors);
+        if ($startingFromEventId === 0) {
+            $projectors->all()->each(function (Projector $projector) {
+                if (method_exists($projector, 'resetState')) {
+                    $projector->resetState();
+                }
+            });
+        }
 
         event(new StartingEventReplay($projectors->all()));
 
         $projectors->call('onStartingEventReplay');
 
-        $this->getStoredEventClass()::query()
-            ->after($afterStoredEventId ?? 0)
-            ->chunk($this->config['replay_chunk_size'], function (Collection $storedEvents) use ($projectors, $onEventReplayed) {
+        $this->storedEventClass::query()
+            ->startingFrom($startingFromEventId ?? 0)
+            ->chunk($this->replayChunkSize, function (Collection $storedEvents) use ($projectors, $onEventReplayed) {
                 $storedEvents->each(function (StoredEvent $storedEvent) use ($projectors, $onEventReplayed) {
                     $this->applyStoredEventToProjectors(
                         $storedEvent,
@@ -229,15 +288,5 @@ class Projectionist
         event(new FinishedEventReplay());
 
         $projectors->call('onFinishedEventReplay');
-    }
-
-    protected function getStoredEventClass(): string
-    {
-        return config('event-projector.stored_event_model');
-    }
-
-    protected function getStoredEventJob(): string
-    {
-        return config('event-projector.stored_event_job');
     }
 }