+++ tests/Console/ListCommandTest.php
@@ -1,29 +1,25 @@
 <?php
 
-namespace Spatie\EventProjector\Console;
+namespace Spatie\EventProjector\Tests\Console;
 
-use Illuminate\Support\Facades\Artisan;
 use Spatie\EventProjector\Tests\TestCase;
 use Spatie\EventProjector\Facades\Projectionist;
+use Spatie\EventProjector\Tests\TestClasses\Reactors\BrokeReactor;
 use Spatie\EventProjector\Tests\TestClasses\Projectors\BalanceProjector;
+use Spatie\EventProjector\Tests\TestClasses\Projectors\MoneyAddedCountProjector;
+use Spatie\EventProjector\Tests\TestClasses\AggregateRoots\Projectors\AccountProjector;
 
-class ListCommandTest extends TestCase
+final class ListCommandTest extends TestCase
 {
     /** @test */
-    public function it_can_list_all_existing_projectors()
+    public function it_can_list_all_registered_projectors_and_reactors()
     {
         Projectionist::addProjector(BalanceProjector::class);
+        Projectionist::addProjector(AccountProjector::class);
+        Projectionist::addProjector(MoneyAddedCountProjector::class);
 
-        Artisan::call('event-projector:list');
+        Projectionist::addReactor(BrokeReactor::class);
 
-        $this->assertSeeInConsoleOutput(BalanceProjector::class);
-    }
-
-    /** @test */
-    public function it_works_when_no_projectors_are_added_to_the_projectionist()
-    {
-        Artisan::call('event-projector:list');
-
-        $this->assertSeeInConsoleOutput('No projectors found.');
+        $this->artisan('event-projector:list')->assertExitCode(0);
     }
 }