+++ tests/TestCase.php
@@ -6,15 +6,20 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Database\Schema\Blueprint;
 use Orchestra\Testbench\TestCase as Orchestra;
+use Spatie\EventProjector\Tests\TestClasses\FakeUuid;
 use Spatie\EventProjector\EventProjectorServiceProvider;
 
 abstract class TestCase extends Orchestra
 {
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
         $this->setUpDatabase();
+
+        FakeUuid::reset();
+
+        $this->artisan('event-projector:clear-event-handlers');
     }
 
     protected function getPackageProviders($app)
@@ -31,6 +36,7 @@ abstract class TestCase extends Orchestra
         Schema::create('accounts', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->nullable();
+            $table->string('uuid')->nullable();
             $table->integer('amount')->default(0);
             $table->integer('addition_count')->default(0);
             $table->timestamps();
@@ -39,15 +45,11 @@ abstract class TestCase extends Orchestra
         Schema::dropIfExists('stored_events');
         include_once __DIR__.'/../stubs/create_stored_events_table.php.stub';
         (new \CreateStoredEventsTable())->up();
-
-        Schema::dropIfExists('projector_statuses');
-        include_once __DIR__.'/../stubs/create_projector_statuses_table.php.stub';
-        (new \CreateProjectorStatusesTable())->up();
     }
 
     protected function assertSeeInConsoleOutput(string $text): self
     {
-        $this->assertContains($text, Artisan::output());
+        $this->assertStringContainsString($text, Artisan::output());
 
         return $this;
     }
@@ -56,5 +58,10 @@ abstract class TestCase extends Orchestra
     {
         config()->set($name, $value);
         (new EventProjectorServiceProvider($this->app))->register();
+    }
+
+    protected function pathToTests(): string
+    {
+        return __DIR__;
     }
 }