+++ tests/TestClasses/Projectors/ProjectorThatInvokesAnObject.php
@@ -4,21 +4,21 @@ namespace Spatie\EventProjector\Tests\TestClasses\Projectors;
 
 use Spatie\EventProjector\Projectors\Projector;
 use Spatie\EventProjector\Projectors\ProjectsEvents;
-use Spatie\EventProjector\Tests\TestClasses\Events\MoneyAdded;
+use Spatie\EventProjector\Tests\TestClasses\Events\MoneyAddedEvent;
 
-class AddMoneyToAccount
+final class AddMoneyToAccount
 {
-    public function __invoke(MoneyAdded $event)
+    public function __invoke(MoneyAddedEvent $event)
     {
         $event->account->addMoney($event->amount);
     }
 }
 
-class ProjectorThatInvokesAnObject implements Projector
+final class ProjectorThatInvokesAnObject implements Projector
 {
     use ProjectsEvents;
 
     protected $handlesEvents = [
-        MoneyAdded::class => AddMoneyToAccount::class,
+        MoneyAddedEvent::class => AddMoneyToAccount::class,
     ];
 }